requirements.txt
streamlit
python-dotenv

langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters

pinecone-client
langchain-pinecone

pypdf
tiktoken
